+++ packages/eslint-plugin-wonder-blocks/src/configs/strict.ts
@@ -1,0 +1,10 @@
+import recommended from "./recommended";
+
+export default {
+    ...recommended,
+    rules: {
+        ...recommended.rules,
+        "@khanacademy/wonder-blocks/no-custom-tab-role": "error",
+        "@khanacademy/wonder-blocks/no-invalid-bodytext-parent": "error",
+    },
+};