+++ packages/wonder-blocks-tabs/src/components/__tests__/tablist.test.tsx
@@ -49,7 +49,12 @@ describe("Tablist", () => {
                 const onBlur = jest.fn();
                 render(
                     <Tablist onBlur={onBlur}>
-                        <button role="tab">Tab</button>
+                        <button
+                            // eslint-disable-next-line @khanacademy/wonder-blocks/no-custom-tab-role -- Explicitly testing an example with an element with role="tab"
+                            role="tab"
+                        >
+                            Tab
+                        </button>
                     </Tablist>,
                 );
                 await userEvent.tab(); // Focus on the tab
@@ -101,6 +106,7 @@ describe("Tablist", () => {
                     <div>
                         <Tablist>
                             <button
+                                // eslint-disable-next-line @khanacademy/wonder-blocks/no-custom-tab-role -- Explicitly testing an example with an element with role="tab"
                                 role="tab"
                                 id={tabId}
                                 aria-controls={panelId}