+++ packages/wonder-blocks-card/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @khanacademy/wonder-blocks-card
 
+## 1.4.8
+
+### Patch Changes
+
+- Updated dependencies [d3d2580]
+    - @khanacademy/wonder-blocks-icon-button@11.2.0
+
+## 1.4.7
+
+### Patch Changes
+
+- Updated dependencies [358eb9b]
+    - @khanacademy/wonder-blocks-tokens@16.1.0
+    - @khanacademy/wonder-blocks-icon-button@11.1.6
+
 ## 1.4.6
 
 ### Patch Changes