+++ packages/wonder-blocks-date-picker/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @khanacademy/wonder-blocks-date-picker
 
+## 1.0.4
+
+### Patch Changes
+
+- @khanacademy/wonder-blocks-modal@8.6.1
+
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies [cd3884c]
+    - @khanacademy/wonder-blocks-modal@8.6.0
+
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies [358eb9b]
+    - @khanacademy/wonder-blocks-tokens@16.1.0
+    - @khanacademy/wonder-blocks-form@7.5.6
+    - @khanacademy/wonder-blocks-icon@5.3.9
+    - @khanacademy/wonder-blocks-modal@8.5.17
+    - @khanacademy/wonder-blocks-styles@0.2.40
+
 ## 1.0.1
 
 ### Patch Changes