+++ packages/wonder-blocks-icon-button/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @khanacademy/wonder-blocks-icon-button
 
+## 11.2.0
+
+### Minor Changes
+
+- d3d2580: Forwards the kind prop on to a data-kind attribute for testing
+
+## 11.1.6
+
+### Patch Changes
+
+- Updated dependencies [358eb9b]
+    - @khanacademy/wonder-blocks-tokens@16.1.0
+    - @khanacademy/wonder-blocks-clickable@8.1.5
+    - @khanacademy/wonder-blocks-icon@5.3.9
+    - @khanacademy/wonder-blocks-styles@0.2.40
+    - @khanacademy/wonder-blocks-typography@4.2.31
+
 ## 11.1.5
 
 ### Patch Changes