+++ packages/wonder-blocks-popover/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @khanacademy/wonder-blocks-popover
 
+## 6.2.0
+
+### Minor Changes
+
+- 6ecb7cc: Add autoUpdate prop to Popover component
+
+## 6.1.57
+
+### Patch Changes
+
+- Updated dependencies [d3d2580]
+    - @khanacademy/wonder-blocks-icon-button@11.2.0
+    - @khanacademy/wonder-blocks-modal@8.6.1
+    - @khanacademy/wonder-blocks-tooltip@4.1.70
+
+## 6.1.56
+
+### Patch Changes
+
+- Updated dependencies [cd3884c]
+    - @khanacademy/wonder-blocks-modal@8.6.0
+    - @khanacademy/wonder-blocks-tooltip@4.1.69
+
+## 6.1.55
+
+### Patch Changes
+
+- Updated dependencies [358eb9b]
+    - @khanacademy/wonder-blocks-tokens@16.1.0
+    - @khanacademy/wonder-blocks-icon-button@11.1.6
+    - @khanacademy/wonder-blocks-modal@8.5.17
+    - @khanacademy/wonder-blocks-styles@0.2.40
+    - @khanacademy/wonder-blocks-tooltip@4.1.68
+    - @khanacademy/wonder-blocks-typography@4.2.31
+
 ## 6.1.54
 
 ### Patch Changes